Code cleanup. Remove unused assets. Move code into appropriate places.
