+++ SongBrowserPlugin/Internals/BeatSaberUI.cs
@@ -331,46 +331,5 @@ namespace SongBrowser.Internals
                 img.color = Color.white;
             }
         }
-
-        public static Texture2D LoadTextureRaw(byte[] file)
-        {
-            if (file.Count() > 0)
-            {
-                Texture2D Tex2D = new Texture2D(2, 2);
-                if (Tex2D.LoadImage(file))
-                    return Tex2D;
-            }
-            return null;
-        }
-
-        public static Texture2D LoadTextureFromResources(string resourcePath)
-        {
-            return LoadTextureRaw(GetResource(Assembly.GetCallingAssembly(), resourcePath));
-        }
-
-        public static Sprite LoadSpriteRaw(byte[] image, float PixelsPerUnit = 100.0f)
-        {
-            return LoadSpriteFromTexture(LoadTextureRaw(image), PixelsPerUnit);
-        }
-
-        public static Sprite LoadSpriteFromTexture(Texture2D SpriteTexture, float PixelsPerUnit = 100.0f)
-        {
-            if (SpriteTexture)
-                return Sprite.Create(SpriteTexture, new Rect(0, 0, SpriteTexture.width, SpriteTexture.height), new Vector2(0, 0), PixelsPerUnit);
-            return null;
-        }
-
-        public static Sprite LoadSpriteFromResources(string resourcePath, float PixelsPerUnit = 100.0f)
-        {
-            return LoadSpriteRaw(GetResource(Assembly.GetCallingAssembly(), resourcePath), PixelsPerUnit);
-        }
-
-        public static byte[] GetResource(Assembly asm, string ResourceName)
-        {
-            System.IO.Stream stream = asm.GetManifestResourceStream(ResourceName);
-            byte[] data = new byte[stream.Length];
-            stream.Read(data, 0, (int)stream.Length);
-            return data;
-        }
     }
 }
